+++ extern/hdusd/usd.cpp
@@ -1,0 +1,80 @@
+/* SPDX-License-Identifier: Apache-2.0
+ * Copyright 2011-2022 Blender Foundation */
+
+#include "usd.h"
+
+namespace hdusd {
+
+std::unique_ptr<pxr::UsdStageCache> stageCache;
+
+static PyObject *stage_export_to_str_func(PyObject * /*self*/, PyObject *args)
+{
+  long stageId;
+  int flatten;
+
+  if (!PyArg_ParseTuple(args, "lp", &stageId, &flatten)) {
+    Py_RETURN_NONE;
+  }
+
+  pxr::UsdStageRefPtr stage = stageCache->Find(pxr::UsdStageCache::Id::FromLongInt(stageId));
+
+  if (!stage) {
+    Py_RETURN_NONE;
+  }
+
+  std::string str;
+  if (flatten) {
+    stage->ExportToString(&str);
+  }
+  else {
+    stage->GetRootLayer()->ExportToString(&str);
+  }
+  return PyUnicode_FromString(str.c_str());
+}
+
+static PyObject *stage_free_func(PyObject * /*self*/, PyObject *args)
+{
+  long stageId;
+
+  if (!PyArg_ParseTuple(args, "l", &stageId)) {
+    Py_RETURN_FALSE;
+  }
+
+  pxr::UsdStageRefPtr stage = stageCache->Find(pxr::UsdStageCache::Id::FromLongInt(stageId));
+
+  if (!stage) {
+    Py_RETURN_FALSE;
+  }
+
+  stageCache->Erase(stage);
+
+  Py_RETURN_TRUE;
+}
+
+static PyMethodDef methods[] = {
+  {"stage_export_to_str", stage_export_to_str_func, METH_VARARGS, ""},
+  {"stage_free", stage_free_func, METH_VARARGS, ""},
+  {NULL, NULL, 0, NULL},
+};
+
+static struct PyModuleDef module = {
+  PyModuleDef_HEAD_INIT,
+  "usd_node",
+  "This module provides access to USD nodes evaluation functions.",
+  -1,
+  methods,
+  NULL,
+  NULL,
+  NULL,
+  NULL,
+};
+
+PyObject *usd_addPythonSubmodule(PyObject *mod)
+{
+  PyObject *submodule = PyModule_Create(&module);
+  PyModule_AddObject(mod, "usd", submodule);
+  return submodule;
+}
+
+}   // namespace hdusd
+
