+++ extern/hdusd/usd_node.h
@@ -1,0 +1,12 @@
+/* SPDX-License-Identifier: Apache-2.0
+ * Copyright 2011-2022 Blender Foundation */
+
+#pragma once
+
+#include <Python.h>
+
+namespace hdusd {
+
+PyObject *usd_node_addPythonSubmodule(PyObject *mod);
+
+} // namespace hdusd
